Remove statusCode from details in Array errors
This is to preserve behavior from strong-remoting rest adapter
