+++ lib/data-builder.js
@@ -5,6 +5,7 @@
 
 'use strict';
 
+var cloneAllProperties = require('../lib/clone.js');
 var httpStatus = require('http-status');
 
 module.exports = function buildResponseData(err, isDebugMode) {
@@ -44,6 +45,7 @@ function serializeArrayOfErrors(errors) {
 
     var data = {};
     cloneAllProperties(data, err);
+    delete data.statusCode;
     details.push(data);
   }
 
@@ -73,17 +75,4 @@ function fillBadRequestError(data, err) {
 
 function fillInternalError(data, err) {
   data.message = httpStatus[data.statusCode] || 'Unknown Error';
-}
-
-function cloneAllProperties(data, err) {
-  // NOTE err.name and err.message are not enumerable properties
-  data.name = err.name;
-  data.message = err.message;
-  for (var p in err) {
-    if ((p in data)) continue;
-    data[p] = err[p];
-  }
-  // NOTE err.stack is not an enumerable property
-  // stack is appended last to ensure order is the same for response
-  data.stack = err.stack;
 }
